requirements.txt
python-dotenv
PyPDF2
Pillow
langchain-core
langchain-community
langchain-google-genai
streamlit
faiss-cpu